Поправил комментарии
