+++ VkNet/Model/Group/GroupInfo.cs
